TST: add image comparison tests with pytest-mpl
