+++ .github/workflows/ci.yml
@@ -1,5 +1,4 @@
 name: CI
-
 on:
   push:
     branches:
@@ -8,6 +7,10 @@ on:
     paths-ignore:
       - README.md
 
+  schedule:
+    # run this once a week (wednesday) at 3 am UTC
+    - cron: '0 3 * * 3'
+  workflow_dispatch:
 jobs:
   tests:
     strategy:
@@ -63,3 +66,50 @@ jobs:
 
     - name: Run mypy
       run: mypy nonos
+
+
+  image-tests:
+    name: Image tests
+    runs-on: ubuntu-latest
+
+    steps:
+    - name: Set up Python
+      uses: actions/setup-python@v4
+      with:
+        python-version: '3.10'
+
+    - uses: actions/checkout@v3
+
+    - name: Build
+      run: python -m pip install ".[dev]"
+
+    - name: Run Image Tests
+      run: |
+        pytest --color=yes --mpl -m mpl_image_compare \
+               --mpl-generate-summary=html \
+               --mpl-results-path=nonos_pytest_mpl_results \
+               --mpl-baseline-path=tests/pytest_mpl_baseline
+
+    - name: Generate new image baseline
+      if: failure()
+      run: |
+        pytest --color=yes --mpl -m mpl_image_compare \
+               --mpl-generate-path=nonos_pytest_mpl_new_baseline \
+               --last-failed
+
+    # always attempt to upload artifacts, even
+    # (and especially) in case of failure.
+    - name: Upload pytest-mpl report
+      if: always()
+      uses: actions/upload-artifact@v2
+      with:
+        name: nonos_pytest_mpl_results
+        path: nonos_pytest_mpl_results/*
+
+    - name: Upload pytest-mpl baseline
+      if: always()
+      uses: actions/upload-artifact@v2
+      with:
+        name: nonos_pytest_mpl_new_baseline
+        path: nonos_pytest_mpl_new_baseline/*
+        if-no-files-found: ignore
