+++ setup.cfg
@@ -38,6 +38,7 @@ dev =
     cogapp>=3.3.0
     memory-profiler
     pytest>=6.1
+    pytest-mpl>=0.15.1
 minimal =
     cblind==2.2.2
     inifix==1.2.0
