+++ tests/test_plotting.py
@@ -1,37 +1,12 @@
 import os
 from glob import glob
-from pathlib import Path
 
-import matplotlib.pyplot as plt
 import numpy.testing as npt
 import pytest
 from matplotlib.colors import SymLogNorm
 
 from nonos.api import GasDataSet, find_nearest
 from nonos.main import main
-
-
-@pytest.fixture()
-def test_data_dir():
-    return Path(__file__).parent / "data"
-
-
-@pytest.fixture(params=["idefix_planet3d", "fargo3d_planet2d"])
-def planet_simulation_dir(test_data_dir, request):
-    return test_data_dir / request.param
-
-
-@pytest.fixture(params=["idefix_rwi", "idefix_planet3d", "fargo3d_planet2d"])
-def simulation_dir(test_data_dir, request):
-    return test_data_dir / request.param
-
-
-@pytest.fixture()
-def temp_figure_and_axis():
-    fig, ax = plt.subplots()
-    yield (fig, ax)
-    plt.close(fig)
-
 
 ARGS_TO_CHECK = {
     "vanilla_conf": ["-geometry", "polar"],
